split ellipses into own file
