+++ apps/editor.planx.uk/src/pages/FlowEditor/components/forms/AddComponentModal/PatternsTab/PatternDetailPanel.tsx
@@ -1,7 +1,7 @@
 import Box from "@mui/material/Box";
 import Stack from "@mui/material/Stack";
 import Typography from "@mui/material/Typography";
-import { EllipsesText } from "components/DelayedLoadingIndicator/DelayedLoadingIndicator";
+import { EllipsesText } from "components/EllipsesText/EllipsesText";
 import React from "react";
 import { FONT_WEIGHT_SEMI_BOLD } from "theme";
